ref(trace-view): Move shared waterfall components out
Currently, trace view and span view share a lot of the same UI components. This
change refactors out many of the shared components into a shared component that
can be imported.
