+++ src/sentry/static/sentry/app/components/waterfallTree/constants.tsx
@@ -1,0 +1,2 @@
+export const ROW_HEIGHT = 24;
+export const ROW_PADDING = 4;
