+++ src/sentry/static/sentry/app/components/waterfallTree/types.tsx
@@ -1,0 +1,1 @@
+export type DurationDisplay = 'left' | 'right' | 'inset';
