+++ src/sentry/static/sentry/app/utils/discover/fieldRenderers.tsx
@@ -5,11 +5,7 @@ import partial from 'lodash/partial';
 
 import Count from 'app/components/count';
 import Duration from 'app/components/duration';
-import {
-  DurationPill,
-  getDurationDisplay,
-  SpanBarRectangle,
-} from 'app/components/events/interfaces/spans/spanBar';
+import {getDurationDisplay} from 'app/components/events/interfaces/spans/spanBar';
 import {
   getHumanDuration,
   pickSpanBarColour,
@@ -19,6 +15,7 @@ import ProjectBadge from 'app/components/idBadge/projectBadge';
 import UserBadge from 'app/components/idBadge/userBadge';
 import UserMisery from 'app/components/userMisery';
 import Version from 'app/components/version';
+import {DurationPill, RowRectangle} from 'app/components/waterfallTree/rowBar';
 import {t} from 'app/locale';
 import {Organization} from 'app/types';
 import {defined} from 'app/utils';
@@ -497,7 +494,7 @@ const spanOperationBreakdownRenderer = (field: string) => (
 
   return (
     <div style={{position: 'relative'}}>
-      <SpanBarRectangle
+      <RowRectangle
         spanBarHatch={false}
         style={{
           backgroundColor: pickSpanBarColour(operationName),
@@ -515,7 +512,7 @@ const spanOperationBreakdownRenderer = (field: string) => (
         >
           {getHumanDuration(spanOpDuration / 1000)}
         </DurationPill>
-      </SpanBarRectangle>
+      </RowRectangle>
     </div>
   );
 };
