+++ src/sentry/static/sentry/app/views/performance/traceDetails/content.tsx
@@ -16,6 +16,7 @@ import Link from 'app/components/links/link';
 import LoadingError from 'app/components/loadingError';
 import LoadingIndicator from 'app/components/loadingIndicator';
 import TimeSince from 'app/components/timeSince';
+import {MessageRow} from 'app/components/waterfallTree/messageRow';
 import {IconInfo} from 'app/icons';
 import {t, tct, tn} from 'app/locale';
 import {Organization} from 'app/types';
@@ -37,7 +38,6 @@ import {
   TraceDetailHeader,
   TraceViewContainer,
   TraceViewHeaderContainer,
-  TransactionRowMessage,
   VirtualScrollBar,
   VirtualScrollBarGrip,
 } from './styles';
@@ -310,7 +310,7 @@ class TraceDetailsContent extends React.Component<Props, State> {
       return null;
     }
 
-    return <TransactionRowMessage>{messages}</TransactionRowMessage>;
+    return <MessageRow>{messages}</MessageRow>;
   }
 
   renderLimitExceededMessage(traceInfo: TraceInfo) {
@@ -325,7 +325,7 @@ class TraceDetailsContent extends React.Component<Props, State> {
     const target = traceEventView.getResultsViewUrlTarget(organization.slug);
 
     return (
-      <TransactionRowMessage>
+      <MessageRow>
         {tct(
           'Limited to a view of [count] transactions. To view the full list, [discover].',
           {
@@ -341,7 +341,7 @@ class TraceDetailsContent extends React.Component<Props, State> {
             ),
           }
         )}
-      </TransactionRowMessage>
+      </MessageRow>
     );
   }
